Refactor portal components to streamline imports and enhance code clarity
- Updated import paths for mutation and query actions across various components to utilize a unified `actions` module.
- Removed unnecessary memoization in several components, simplifying the code structure.
- Refactored event handler functions to eliminate the use of `useCallback` where not needed, improving readability.

These changes contribute to a cleaner and more maintainable codebase, ensuring consistent import practices across the portal components.
