+++ src/app/_components/PortalComponents/DeleteClerkUserCell/index.tsx
@@ -1,7 +1,7 @@
 import { useState, useTransition } from "react";
 
 import styles from "@/app/_components/PortalComponents/DeleteClerkUserCell/index.module.scss";
-import { deleteClerkUser } from "@/app/portal/_actions/mutations";
+import { deleteClerkUser } from "@/app/portal/actions";
 import { runPortalAction } from "@/app/portal/_lib/runAction";
 
 import BasicButton from "../../Buttons/BasicButton";
