+++ src/app/_components/PortalComponents/DropZone/index.tsx
@@ -1,5 +1,5 @@
 import Image, { type StaticImageData } from "next/image";
-import React, { useCallback } from "react";
+import React from "react";
 
 import Upload from "../../svgs/Upload";
 import styles from "./index.module.scss";
@@ -10,44 +10,38 @@ interface DropZoneProps {
 }
 
 const DropZone = ({ currentImage, handleFileUpload }: DropZoneProps) => {
-  const handleDrop = useCallback(
-    (e: React.DragEvent) => {
-      e.preventDefault();
-      const files = Array.from(e.dataTransfer.files)
-        .filter(
-          (file) =>
-            file.type.startsWith("image/") ||
-            file.name.toLowerCase().endsWith("jxl"),
-        )
-        .map((file) => ({
-          file,
-        }));
-      if (files[0]?.file) {
-        handleFileUpload(files[0]?.file);
-      }
-    },
-    [handleFileUpload],
-  );
-
-  const handleDragOver = useCallback((e: React.DragEvent) => {
+  const handleDrop = (e: React.DragEvent) => {
     e.preventDefault();
-  }, []);
+    const files = Array.from(e.dataTransfer.files)
+      .filter(
+        (file) =>
+          file.type.startsWith("image/") ||
+          file.name.toLowerCase().endsWith("jxl"),
+      )
+      .map((file) => ({
+        file,
+      }));
+    if (files[0]?.file) {
+      handleFileUpload(files[0]?.file);
+    }
+  };
 
-  const handleFileInput = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const files = Array.from(e.target.files ?? [])
-        .filter((file) => file.type.startsWith("image/"))
-        .map((file) => ({
-          file,
-          id: crypto.randomUUID(),
-        }));
-      if (files[0]?.file) {
-        handleFileUpload(files[0]?.file);
-      }
-      e.target.value = "";
-    },
-    [handleFileUpload],
-  );
+  const handleDragOver = (e: React.DragEvent) => {
+    e.preventDefault();
+  };
+
+  const handleFileInput = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const files = Array.from(e.target.files ?? [])
+      .filter((file) => file.type.startsWith("image/"))
+      .map((file) => ({
+        file,
+        id: crypto.randomUUID(),
+      }));
+    if (files[0]?.file) {
+      handleFileUpload(files[0]?.file);
+    }
+    e.target.value = "";
+  };
 
   return (
     <div
