+++ src/app/_components/PortalComponents/EditRecruitmentFormCell/DeleteForm.tsx
@@ -1,7 +1,7 @@
 import { useTransition } from "react";
 
 import styles from "@/app/_components/PortalComponents/EditSponsorCell/index.module.scss";
-import { deleteRecruitmentForm } from "@/app/portal/_actions/mutations";
+import { deleteRecruitmentForm } from "@/app/portal/actions";
 import { runPortalAction } from "@/app/portal/_lib/runAction";
 import { type UserResource } from "@clerk/nextjs/types";
