+++ src/app/_components/PortalComponents/EditSponsorCell/DeleteSponsor.tsx
@@ -1,7 +1,7 @@
 import { useTransition } from "react";
 
 import styles from "@/app/_components/PortalComponents/EditSponsorCell/index.module.scss";
-import { deleteSponsor } from "@/app/portal/_actions/mutations";
+import { deleteSponsor } from "@/app/portal/actions";
 import { runPortalAction } from "@/app/portal/_lib/runAction";
 
 import BasicButton, { ButtonVariant } from "../../Buttons/BasicButton";
