+++ src/app/_components/PortalComponents/EditUserCell/MoveToAlumniModal.tsx
@@ -3,7 +3,7 @@
 import { useState, useTransition } from "react";
 import toast from "react-hot-toast";
 
-import { moveUserToAlumni } from "@/app/portal/_actions/mutations";
+import { moveUserToAlumni } from "@/app/portal/actions";
 import { runPortalAction } from "@/app/portal/_lib/runAction";
 
 import styles from "./index.module.scss";
