+++ src/app/_components/PortalComponents/Portal/Invitations/RevokeUserCell.tsx
@@ -2,7 +2,7 @@ import { useTransition } from "react";
 
 import BasicButton from "@/app/_components/Buttons/BasicButton";
 import styles from "@/app/_components/PortalComponents/DeleteClerkUserCell/index.module.scss";
-import { revokeUserInvitation } from "@/app/portal/_actions/mutations";
+import { revokeUserInvitation } from "@/app/portal/actions";
 import { runPortalAction } from "@/app/portal/_lib/runAction";
 
 export interface RevokeUserInvitationProps {
