+++ src/app/portal/(portal)/(admin)/invitations/page.tsx
@@ -1,5 +1,5 @@
 import InvitationsTable from "@/app/_components/PortalComponents/Portal/Invitations/InvitationsTable";
-import { getInvitedUsers } from "@/app/portal/_actions/queries";
+import { getInvitedUsers } from "@/app/portal/actions";
 
 export const metadata = {
   title: "Invitations - Portal",
