+++ src/app/portal/(portal)/(admin)/recruitment/page.tsx
@@ -1,5 +1,5 @@
 import RecruitmentTable from "@/app/_components/PortalComponents/Portal/recruitment/RecruitmentTable";
-import { getFormsList } from "@/app/portal/_actions/queries";
+import { getFormsList } from "@/app/portal/actions";
 
 export const metadata = {
   title: "Recruitment - Portal",
