+++ src/app/portal/(portal)/(admin)/users/page.tsx
@@ -1,5 +1,5 @@
 import UsersTable from "@/app/_components/PortalComponents/Portal/users/UsersTable";
-import { getClerkUsers } from "@/app/portal/_actions/queries";
+import { getClerkUsers } from "@/app/portal/actions";
 
 export const metadata = {
   title: "Users - Portal",
